Encapsulate jwt auth service
